+++ tools/themes/tool.js
@@ -8,7 +8,7 @@
         name:'themes',
         category: 'general',
         type:'multiple',
-        possibleValues: ['performance','justExplorer','mobile','slimArrows','flairX','lessIcons','nonStickyHeader','noStudyChat','pieceDrag'],
+        possibleValues: ['performance','justExplorer','mobile','slimArrows','flairX','lessIcons','nonStickyHeader','noStudyChat','pieceDrag','noGrab'],
         defaultValue: '',
         advanced: true
       }
@@ -27,7 +27,8 @@
         'themes.lessIcons':'Less Icons',
         'themes.nonStickyHeader':'No sticky header',
         'themes.noStudyChat':'No study chat',
-        'themes.pieceDrag':'Nicer piece drag'
+        'themes.pieceDrag':'Nicer piece drag',
+        'themes.noGrab':'No grab cursor'
       },
       'ro-RO':{
         'options.general': 'General',
@@ -41,7 +42,8 @@
         'themes.lessIcons':'Mai pu\u0163ine pictograme',
         'themes.nonStickyHeader':'Antet fix \u00een pagin\u0103',
         'themes.noStudyChat':'F\u0103r\u0103 chat \u00eeen studii',
-        'themes.pieceDrag':'Apucat piese mai frumos'
+        'themes.pieceDrag':'Apucat piese mai frumos',
+        'themes.noGrab':'F\u0103r\u0103 cursor care apuc\u0103'
       }
     }