+++ HISTORY.md
@@ -1,5 +1,8 @@
 History of features added to LiChess Tools in time (reversed order)
 
+29 May
+  - No hand grab cursor theme
+
 27 May
   - Wakelock for TV pages